requirements.txt
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0
requests>=2.31.0
streamlit==1.35.0
plotly==5.22.0
python-dotenv==1.0.1
